+++ services/api/pepc_boost_test.go
@@ -1,102 +1,31 @@
 package api
 
 import (
+	"context"
+	"fmt"
 	"math/big"
-	"strconv"
+	"net/http"
 	"testing"
 
-	"github.com/alicebob/miniredis/v2"
+	"github.com/attestantio/go-eth2-client/spec/bellatrix"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
+	bellatrixUtil "github.com/attestantio/go-eth2-client/util/bellatrix"
+	common2 "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	gethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/flashbots/go-boost-utils/bls"
-	boostTypes "github.com/flashbots/go-boost-utils/types"
+	"github.com/flashbots/go-boost-utils/types"
+	boosttypes "github.com/flashbots/go-boost-utils/types"
 	"github.com/flashbots/mev-boost-relay/beaconclient"
 	"github.com/flashbots/mev-boost-relay/common"
-	"github.com/flashbots/mev-boost-relay/database"
-	"github.com/flashbots/mev-boost-relay/datastore"
 	"github.com/stretchr/testify/require"
 )
 
-func startTestBackend(t *testing.T) (*phase0.BLSPubKey, *bls.SecretKey, *testBackend) {
-	t.Helper()
-	// Setup test key pair.
-	sk, _, err := bls.GenerateNewKeypair()
-	require.NoError(t, err)
-	blsPubkey, err := bls.PublicKeyFromSecretKey(sk)
-	require.NoError(t, err)
-	pkBytes := blsPubkey.Bytes()
-	var pubkey phase0.BLSPubKey
-	copy(pubkey[:], pkBytes[:])
-	pkStr := pubkey.String()
-
-	// Setup test backend.
-	backend := newTestBackend(t, 1)
-	backend.relay.genesisInfo = &beaconclient.GetGenesisResponse{}
-	backend.relay.genesisInfo.Data.GenesisTime = 0
-	backend.relay.proposerDutiesMap = map[uint64]*common.BuilderGetValidatorsResponseEntry{
-		slot: {
-			Entry: &boostTypes.SignedValidatorRegistration{
-				Message: &boostTypes.RegisterValidatorRequestMessage{
-					FeeRecipient: [20]byte(feeRecipient),
-					GasLimit:     5000,
-					Timestamp:    0xffffffff,
-					Pubkey:       [48]byte(phase0.BLSPubKey{}),
-				},
-			},
-		},
-	}
-	backend.relay.opts.BlockBuilderAPI = true
-	backend.relay.beaconClient = beaconclient.NewMockMultiBeaconClient()
-	backend.relay.blockSimRateLimiter = &MockBlockSimulationRateLimiter{}
-	backend.relay.blockBuildersCache = map[string]*blockBuilderCacheEntry{
-		pkStr: {
-			status: common.BuilderStatus{
-				IsHighPrio:   true,
-				IsOptimistic: true,
-			},
-			collateral: big.NewInt(int64(collateral)),
-		},
-	}
-
-	// Setup test db, redis, and datastore.
-	mockDB := &database.MockDB{
-		Builders: map[string]*database.BlockBuilderEntry{
-			pkStr: {
-				BuilderPubkey: pkStr,
-				IsHighPrio:    true,
-				IsOptimistic:  true,
-				BuilderID:     builderID,
-				Collateral:    strconv.Itoa(collateral),
-			},
-		},
-		Demotions: map[string]bool{},
-		Refunds:   map[string]bool{},
-	}
-	redisTestServer, err := miniredis.Run()
-	require.NoError(t, err)
-	mockRedis, err := datastore.NewRedisCache("", redisTestServer.Addr(), "")
-	require.NoError(t, err)
-	mockDS, err := datastore.NewDatastore(mockRedis, nil, mockDB)
-	require.NoError(t, err)
-
-	backend.relay.datastore = mockDS
-	backend.relay.redis = mockRedis
-	backend.relay.db = mockDB
-
-	// Prepare redis
-	// err = backend.relay.redis.SetKnownValidator(boostTypes.NewPubkeyHex(pubkey.String()), proposerInd)
-	// require.NoError(t, err)
-
-	// count, err := backend.relay.datastore.RefreshKnownValidators()
-	// require.NoError(t, err)
-	// require.Equal(t, count, 1)
-
-	backend.relay.headSlot.Store(40)
-	return &pubkey, sk, backend
-}
-
 // TODO - this test will keep evolving as we expand the state interference checks
-func TestCcheckTxAndSenderValidity(t *testing.T) {
+func TestCheckTxAndSenderValidity(t *testing.T) {
+	_, _, backend := startTestBackend(t)
+	randomAddress := common2.BytesToAddress([]byte("0xabc"))
+	uniswapV2Address := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
 	cases := []struct {
 		description   string
 		txs           []*gethtypes.Transaction
@@ -105,15 +34,1005 @@ func TestCcheckTxAndSenderValidity(t *testing.T) {
 		{
 			description:   "no txs sent",
 			txs:           []*gethtypes.Transaction{},
-			requiredError: "empty txs sent",
+			requiredError: "Empty TOB tx request sent!",
+		},
+		{
+			description: "only 1 tx sent",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(2),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "We require a payment tx along with the TOB txs!",
+		},
+		{
+			description: "payout not addresses to relayer",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &uniswapV2Address,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &randomAddress,
+					Value:    big.NewInt(2),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "We require a payment tx to the relayer along with the TOB txs!",
+		},
+		{
+			description: "zero value payout",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &uniswapV2Address,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(0),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "The relayer payment tx is non-zero!",
+		},
+		{
+			description: "malformed payout",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &uniswapV2Address,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(110),
+					Data:     []byte("tx2"),
+				}),
+			},
+			requiredError: "The relayer payment tx has malformed data!",
+		},
+		{
+			description: "More than 2 txs sent",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       nil,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       nil,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx2"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(110),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "we support only 1 tx on the TOB currently, got 3",
+		},
+		{
+			description: "First tx is a contract creation",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       nil,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(110),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "contract creation cannot be a TOB tx",
+		},
+		{
+			description: "ToB tx is not part of whitelist",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &randomAddress,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(110),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "TOB tx can only be sent to uniswap v2 router",
+		},
+		{
+			description: "Valid ToB txs",
+			txs: []*gethtypes.Transaction{
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &uniswapV2Address,
+					Value:    big.NewInt(2),
+					Data:     []byte("tx1"),
+				}),
+				gethtypes.NewTx(&gethtypes.LegacyTx{
+					Nonce:    2,
+					GasPrice: big.NewInt(2),
+					Gas:      2,
+					To:       &backend.relay.relayerPayoutAddress,
+					Value:    big.NewInt(110),
+					Data:     []byte(""),
+				}),
+			},
+			requiredError: "",
 		},
 	}
 
 	for _, c := range cases {
 		t.Run(c.description, func(t *testing.T) {
-			_, _, backend := startTestBackend(t)
-			backend.relay.checkTxAndSenderValidity(c.txs)
-
+			err := backend.relay.checkTxAndSenderValidity(c.txs)
+			if c.requiredError != "" {
+				require.Contains(t, err.Error(), c.requiredError)
+			} else {
+				require.NoError(t, err)
+			}
 		})
 	}
+}
+
+// TODO - refactor these tests into a TDT
+func TestSubmitTobTxsOverrideTxsWithHigherValue(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 1 : Happy path
+	req := new(common.TobTxsSubmitRequest)
+
+	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    1,
+		GasPrice: big.NewInt(1),
+		Gas:      1,
+		To:       &addr1,
+		Value:    big.NewInt(1),
+		Data:     []byte("tx1"),
+	})
+	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    2,
+		GasPrice: big.NewInt(2),
+		Gas:      2,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(2),
+		Data:     []byte(""),
+	})
+	tx1byte, err := tx1.MarshalBinary()
+	require.NoError(t, err)
+	tx2byte, err := tx2.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusOK, rr.Code)
+
+	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
+	require.Equal(t, tobTxValue, big.NewInt(2))
+
+	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+
+	require.Equal(t, 2, len(tobtxs))
+
+	firstTx := new(gethtypes.Transaction)
+	err = firstTx.UnmarshalBinary(tobtxs[0])
+
+	firstTxJson, err := firstTx.MarshalJSON()
+	require.NoError(t, err)
+	tx1Json, err := tx1.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, firstTxJson, tx1Json)
+
+	secondTx := new(gethtypes.Transaction)
+	err = secondTx.UnmarshalBinary(tobtxs[1])
+	secondTxJson, err := secondTx.MarshalJSON()
+	require.NoError(t, err)
+	tx2Json, err := tx2.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, secondTxJson, tx2Json)
+
+	// Test 2: Try adding txs with higher value
+	req = new(common.TobTxsSubmitRequest)
+	addr1 = common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx3 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx3"),
+	})
+	tx4 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(10),
+		Data:     []byte(""),
+	})
+	tx3byte, err := tx3.MarshalBinary()
+	require.NoError(t, err)
+	tx4byte, err := tx4.MarshalBinary()
+	require.NoError(t, err)
+	txs = bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx3byte, tx4byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err = req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusOK, rr.Code)
+
+	tobTxValue, err = backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
+	require.Equal(t, tobTxValue, big.NewInt(10))
+
+	tobtxs, err = backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+
+	require.Equal(t, 2, len(tobtxs))
+
+	firstTx = new(gethtypes.Transaction)
+	err = firstTx.UnmarshalBinary(tobtxs[0])
+
+	firstTxJson, err = firstTx.MarshalJSON()
+	require.NoError(t, err)
+	tx1Json, err = tx3.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, firstTxJson, tx1Json)
+
+	secondTx = new(gethtypes.Transaction)
+	err = secondTx.UnmarshalBinary(tobtxs[1])
+	secondTxJson, err = secondTx.MarshalJSON()
+	require.NoError(t, err)
+	tx2Json, err = tx4.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, secondTxJson, tx2Json)
+}
+
+func TestSubmitTobTxsLowerValueRequests(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 1 : Happy path
+	req := new(common.TobTxsSubmitRequest)
+	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    1,
+		GasPrice: big.NewInt(1),
+		Gas:      1,
+		To:       &addr1,
+		Value:    big.NewInt(1),
+		Data:     []byte("tx1"),
+	})
+	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    2,
+		GasPrice: big.NewInt(2),
+		Gas:      2,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(2),
+		Data:     []byte(""),
+	})
+	tx1byte, err := tx1.MarshalBinary()
+	require.NoError(t, err)
+	tx2byte, err := tx2.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusOK, rr.Code)
+
+	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
+	require.Equal(t, tobTxValue, big.NewInt(2))
+
+	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+
+	require.Equal(t, 2, len(tobtxs))
+
+	firstTx := new(gethtypes.Transaction)
+	err = firstTx.UnmarshalBinary(tobtxs[0])
+
+	firstTxJson, err := firstTx.MarshalJSON()
+	require.NoError(t, err)
+	tx1Json, err := tx1.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, firstTxJson, tx1Json)
+
+	secondTx := new(gethtypes.Transaction)
+	err = secondTx.UnmarshalBinary(tobtxs[1])
+	secondTxJson, err := secondTx.MarshalJSON()
+	require.NoError(t, err)
+	tx2Json, err := tx2.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, secondTxJson, tx2Json)
+
+	// Test 2: Try adding txs with higher value
+	req = new(common.TobTxsSubmitRequest)
+	addr1 = common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx3 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx3"),
+	})
+	tx4 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(10),
+		Data:     []byte(""),
+	})
+	tx3byte, err := tx3.MarshalBinary()
+	require.NoError(t, err)
+	tx4byte, err := tx4.MarshalBinary()
+	require.NoError(t, err)
+	txs = bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx3byte, tx4byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err = req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusOK, rr.Code)
+
+	tobTxValue, err = backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
+	require.Equal(t, tobTxValue, big.NewInt(10))
+
+	tobtxs, err = backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+
+	require.Equal(t, 2, len(tobtxs))
+
+	firstTx = new(gethtypes.Transaction)
+	err = firstTx.UnmarshalBinary(tobtxs[0])
+
+	firstTxJson, err = firstTx.MarshalJSON()
+	require.NoError(t, err)
+	tx1Json, err = tx3.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, firstTxJson, tx1Json)
+
+	secondTx = new(gethtypes.Transaction)
+	err = secondTx.UnmarshalBinary(tobtxs[1])
+	secondTxJson, err = secondTx.MarshalJSON()
+	require.NoError(t, err)
+	tx2Json, err = tx4.MarshalJSON()
+	require.NoError(t, err)
+	require.Equal(t, secondTxJson, tx2Json)
+}
+
+func TestSubmitTobTxsNoPayout(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 3: No payout tx
+	req := new(common.TobTxsSubmitRequest)
+	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx7 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx6"),
+	})
+	tx7byte, err := tx7.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx7byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusBadRequest, rr.Code)
+	require.Contains(t, rr.Body.String(), "We require a payment tx along with the TOB txs!")
+}
+
+func TestSubmitTobTxsPayoutToWrongAddress(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 3: Payout tx is to the wrong address
+	req := new(common.TobTxsSubmitRequest)
+	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx8 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx6"),
+	})
+	tx9 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &addr1,
+		Value:    big.NewInt(5),
+		Data:     []byte(""),
+	})
+	tx8byte, err := tx8.MarshalBinary()
+	require.NoError(t, err)
+	tx9byte, err := tx9.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx8byte, tx9byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusBadRequest, rr.Code)
+	require.Contains(t, rr.Body.String(), "We require a payment tx to the relayer along with the TOB txs!")
+
+}
+
+func TestSubmitTobTxsStateInterference(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 3: Fail state interference checks
+	req := new(common.TobTxsSubmitRequest)
+	addr1 := common2.HexToAddress("0xB2D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx6"),
+	})
+	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(5),
+		Data:     []byte(""),
+	})
+	tx1byte, err := tx1.MarshalBinary()
+	require.NoError(t, err)
+	tx2byte, err := tx2.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusBadRequest, rr.Code)
+	require.Contains(t, rr.Body.String(), "TOB tx can only be sent to uniswap v2 router")
+
+}
+
+func TestSubmitTobTxsCheckSlots(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	// Test 4 : Slot to far ahead
+	addr1 := common2.HexToAddress("0xB2D7a3554F221B34f49d7d3C61375E603aFb699e")
+	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx6"),
+	})
+	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(5),
+		Data:     []byte(""),
+	})
+	tx1byte, err := tx1.MarshalBinary()
+	require.NoError(t, err)
+	tx2byte, err := tx2.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
+
+	req := &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 2,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusBadRequest, rr.Code)
+	require.Contains(t, rr.Body.String(), "Slot's TOB bid not yet started!!")
+
+	// Test 5: No txs are sent
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{}},
+		Slot:       headSlot + 2,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err = req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusBadRequest, rr.Code)
+	require.Contains(t, rr.Body.String(), "Empty TOB tx request sent")
+}
+
+func TestSubmitTobTxsHappyPath(t *testing.T) {
+	path := "/relay/v1/builder/tob_txs"
+	backend := newTestBackend(t, 1)
+
+	headSlot := uint64(32)
+	submissionSlot := headSlot + 1
+	//submissionTimestamp := 1606824419
+
+	// Payload attributes
+	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
+	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
+	require.NoError(t, err)
+	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
+	require.NoError(t, err)
+	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
+	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
+	require.NoError(t, err)
+	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
+
+	// Setup the test relay backend
+	backend.relay.headSlot.Store(headSlot)
+	backend.relay.capellaEpoch = 1
+	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
+	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
+		Slot: headSlot,
+		Entry: &types.SignedValidatorRegistration{
+			Message: &types.RegisterValidatorRequestMessage{
+				Pubkey:       boosttypes.PublicKey(proposerPubkey),
+				FeeRecipient: feeRec,
+			},
+		},
+	}
+	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
+	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
+		slot:       submissionSlot,
+		parentHash: parentHash,
+		payloadAttributes: beaconclient.PayloadAttributes{
+			PrevRandao: prevRandao,
+		},
+		withdrawalsRoot: phase0.Root(withdrawalsRoot),
+	}
+
+	req := new(common.TobTxsSubmitRequest)
+	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
+
+	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    3,
+		GasPrice: big.NewInt(3),
+		Gas:      3,
+		To:       &addr1,
+		Value:    big.NewInt(3),
+		Data:     []byte("tx3"),
+	})
+	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
+		Nonce:    4,
+		GasPrice: big.NewInt(5),
+		Gas:      12,
+		To:       &backend.relay.relayerPayoutAddress,
+		Value:    big.NewInt(10),
+		Data:     []byte(""),
+	})
+	tx1byte, err := tx1.MarshalBinary()
+	require.NoError(t, err)
+	tx2byte, err := tx2.MarshalBinary()
+	require.NoError(t, err)
+	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
+	txsHashRoot, err := txs.HashTreeRoot()
+	require.NoError(t, err)
+	req = &common.TobTxsSubmitRequest{
+		ParentHash: parentHash,
+		TobTxs:     txs,
+		Slot:       headSlot + 1,
+	}
+	fmt.Printf("Marshalling request to json!!")
+	jsonReq, err := req.MarshalJSON()
+	require.NoError(t, err)
+	fmt.Printf("Unmarshalling request from json!!")
+
+	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
+		"Content-Type": "application/json",
+	})
+	require.Equal(t, http.StatusOK, rr.Code)
+
+	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
+	require.Equal(t, tobTxValue, big.NewInt(10))
+
+	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
+	require.NoError(t, err)
+
+	require.Equal(t, 2, len(tobtxs))
+
+	firstTx := new(gethtypes.Transaction)
+	err = firstTx.UnmarshalBinary(tobtxs[0])
+	require.NoError(t, err)
+
+	secondTx := new(gethtypes.Transaction)
+	err = secondTx.UnmarshalBinary(tobtxs[1])
+	require.NoError(t, err)
+
+	firstTxBytes, err := firstTx.MarshalBinary()
+	require.NoError(t, err)
+	secondTxBytes, err := secondTx.MarshalBinary()
+	require.NoError(t, err)
+
+	txsPostStoringInRedis := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{firstTxBytes, secondTxBytes}}
+	txsPostStoringInRedisHashRoot, err := txsPostStoringInRedis.HashTreeRoot()
+	require.NoError(t, err)
+	require.Equal(t, txsHashRoot, txsPostStoringInRedisHashRoot)
 }
