+++ services/api/service_test.go
@@ -18,13 +18,9 @@ import (
 	"github.com/attestantio/go-eth2-client/spec/bellatrix"
 	"github.com/attestantio/go-eth2-client/spec/capella"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
-	bellatrixUtil "github.com/attestantio/go-eth2-client/util/bellatrix"
-	common2 "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	gethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/flashbots/go-boost-utils/bls"
 	"github.com/flashbots/go-boost-utils/types"
-	boosttypes "github.com/flashbots/go-boost-utils/types"
 	"github.com/flashbots/mev-boost-relay/beaconclient"
 	"github.com/flashbots/mev-boost-relay/common"
 	"github.com/flashbots/mev-boost-relay/database"
@@ -873,830 +869,4 @@ func gzipBytes(t *testing.T, b []byte) []byte {
 	require.NoError(t, err)
 	require.NoError(t, zw.Close())
 	return buf.Bytes()
-}
-
-func TestSubmitTobTxsOverrideTxsWithHigherValue(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 1 : Happy path
-	req := new(common.TobTxsSubmitRequest)
-
-	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    1,
-		GasPrice: big.NewInt(1),
-		Gas:      1,
-		To:       &addr1,
-		Value:    big.NewInt(1),
-		Data:     []byte("tx1"),
-	})
-	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    2,
-		GasPrice: big.NewInt(2),
-		Gas:      2,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(2),
-		Data:     []byte(""),
-	})
-	tx1byte, err := tx1.MarshalBinary()
-	require.NoError(t, err)
-	tx2byte, err := tx2.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
-	require.Equal(t, tobTxValue, big.NewInt(2))
-
-	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-
-	require.Equal(t, 2, len(tobtxs))
-
-	firstTx := new(gethtypes.Transaction)
-	err = firstTx.UnmarshalBinary(tobtxs[0])
-
-	firstTxJson, err := firstTx.MarshalJSON()
-	require.NoError(t, err)
-	tx1Json, err := tx1.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, firstTxJson, tx1Json)
-
-	secondTx := new(gethtypes.Transaction)
-	err = secondTx.UnmarshalBinary(tobtxs[1])
-	secondTxJson, err := secondTx.MarshalJSON()
-	require.NoError(t, err)
-	tx2Json, err := tx2.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, secondTxJson, tx2Json)
-
-	// Test 2: Try adding txs with higher value
-	req = new(common.TobTxsSubmitRequest)
-	addr1 = common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx3 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx3"),
-	})
-	tx4 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(10),
-		Data:     []byte(""),
-	})
-	tx3byte, err := tx3.MarshalBinary()
-	require.NoError(t, err)
-	tx4byte, err := tx4.MarshalBinary()
-	require.NoError(t, err)
-	txs = bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx3byte, tx4byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err = req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	tobTxValue, err = backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
-	require.Equal(t, tobTxValue, big.NewInt(10))
-
-	tobtxs, err = backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-
-	require.Equal(t, 2, len(tobtxs))
-
-	firstTx = new(gethtypes.Transaction)
-	err = firstTx.UnmarshalBinary(tobtxs[0])
-
-	firstTxJson, err = firstTx.MarshalJSON()
-	require.NoError(t, err)
-	tx1Json, err = tx3.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, firstTxJson, tx1Json)
-
-	secondTx = new(gethtypes.Transaction)
-	err = secondTx.UnmarshalBinary(tobtxs[1])
-	secondTxJson, err = secondTx.MarshalJSON()
-	require.NoError(t, err)
-	tx2Json, err = tx4.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, secondTxJson, tx2Json)
-}
-
-func TestSubmitTobTxsLowerValueRequests(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 1 : Happy path
-	req := new(common.TobTxsSubmitRequest)
-	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx1 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    1,
-		GasPrice: big.NewInt(1),
-		Gas:      1,
-		To:       &addr1,
-		Value:    big.NewInt(1),
-		Data:     []byte("tx1"),
-	})
-	tx2 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    2,
-		GasPrice: big.NewInt(2),
-		Gas:      2,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(2),
-		Data:     []byte(""),
-	})
-	tx1byte, err := tx1.MarshalBinary()
-	require.NoError(t, err)
-	tx2byte, err := tx2.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx1byte, tx2byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
-	require.Equal(t, tobTxValue, big.NewInt(2))
-
-	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-
-	require.Equal(t, 2, len(tobtxs))
-
-	firstTx := new(gethtypes.Transaction)
-	err = firstTx.UnmarshalBinary(tobtxs[0])
-
-	firstTxJson, err := firstTx.MarshalJSON()
-	require.NoError(t, err)
-	tx1Json, err := tx1.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, firstTxJson, tx1Json)
-
-	secondTx := new(gethtypes.Transaction)
-	err = secondTx.UnmarshalBinary(tobtxs[1])
-	secondTxJson, err := secondTx.MarshalJSON()
-	require.NoError(t, err)
-	tx2Json, err := tx2.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, secondTxJson, tx2Json)
-
-	// Test 2: Try adding txs with higher value
-	req = new(common.TobTxsSubmitRequest)
-	addr1 = common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx3 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx3"),
-	})
-	tx4 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(10),
-		Data:     []byte(""),
-	})
-	tx3byte, err := tx3.MarshalBinary()
-	require.NoError(t, err)
-	tx4byte, err := tx4.MarshalBinary()
-	require.NoError(t, err)
-	txs = bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx3byte, tx4byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err = req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	tobTxValue, err = backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
-	require.Equal(t, tobTxValue, big.NewInt(10))
-
-	tobtxs, err = backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-
-	require.Equal(t, 2, len(tobtxs))
-
-	firstTx = new(gethtypes.Transaction)
-	err = firstTx.UnmarshalBinary(tobtxs[0])
-
-	firstTxJson, err = firstTx.MarshalJSON()
-	require.NoError(t, err)
-	tx1Json, err = tx3.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, firstTxJson, tx1Json)
-
-	secondTx = new(gethtypes.Transaction)
-	err = secondTx.UnmarshalBinary(tobtxs[1])
-	secondTxJson, err = secondTx.MarshalJSON()
-	require.NoError(t, err)
-	tx2Json, err = tx4.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, secondTxJson, tx2Json)
-}
-
-func TestSubmitTobTxsNoPayout(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 3: No payout tx
-	req := new(common.TobTxsSubmitRequest)
-	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx7 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx6"),
-	})
-	tx7byte, err := tx7.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx7byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "We require a payment tx along with the TOB txs!")
-}
-
-func TestSubmitTobTxsPayoutToWrongAddress(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 3: Payout tx is to the wrong address
-	req := new(common.TobTxsSubmitRequest)
-	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx8 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx6"),
-	})
-	tx9 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &addr1,
-		Value:    big.NewInt(5),
-		Data:     []byte(""),
-	})
-	tx8byte, err := tx8.MarshalBinary()
-	require.NoError(t, err)
-	tx9byte, err := tx9.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx8byte, tx9byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "We require a payment tx to the relayer along with the TOB txs!")
-
-}
-
-func TestSubmitTobTxsStateInterference(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 3: Fail state interference checks
-	req := new(common.TobTxsSubmitRequest)
-	addr1 := common2.HexToAddress("0xB2D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx10 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx6"),
-	})
-	tx11 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(5),
-		Data:     []byte(""),
-	})
-	tx10byte, err := tx10.MarshalBinary()
-	require.NoError(t, err)
-	tx11byte, err := tx11.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx10byte, tx11byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "TOB tx can only be sent to uniswap v2 router")
-
-}
-
-func TestSubmitTobTxsCheckSlots(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 4 : Slot to far ahead
-	addr1 := common2.HexToAddress("0xB2D7a3554F221B34f49d7d3C61375E603aFb699e")
-	tx10 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx6"),
-	})
-	tx11 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(5),
-		Data:     []byte(""),
-	})
-	tx10byte, err := tx10.MarshalBinary()
-	require.NoError(t, err)
-	tx11byte, err := tx11.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx10byte, tx11byte}}
-
-	req := &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 2,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "Slot's TOB bid not yet started!!")
-
-	// Test 5: No txs are sent
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{}},
-		Slot:       headSlot + 2,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err = req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "Empty TOB tx request sent")
-}
-
-func TestSubmitTobTxsHappyPath(t *testing.T) {
-	path := "/relay/v1/builder/tob_txs"
-	backend := newTestBackend(t, 1)
-
-	headSlot := uint64(32)
-	submissionSlot := headSlot + 1
-	//submissionTimestamp := 1606824419
-
-	// Payload attributes
-	parentHash := "0xbd3291854dc822b7ec585925cda0e18f06af28fa2886e15f52d52dd4b6f94ed6"
-	feeRec, err := types.HexToAddress("0x5cc0dde14e7256340cc820415a6022a7d1c93a35")
-	require.NoError(t, err)
-	withdrawalsRoot, err := hexutil.Decode("0xb15ed76298ff84a586b1d875df08b6676c98dfe9c7cd73fab88450348d8e70c8")
-	require.NoError(t, err)
-	prevRandao := "0x9962816e9d0a39fd4c80935338a741dc916d1545694e41eb5a505e1a3098f9e4"
-	proposerPubkeyByte, err := hexutil.Decode(testProposerKey)
-	require.NoError(t, err)
-	proposerPubkey := phase0.BLSPubKey(proposerPubkeyByte)
-
-	// Setup the test relay backend
-	backend.relay.headSlot.Store(headSlot)
-	backend.relay.capellaEpoch = 1
-	backend.relay.proposerDutiesMap = make(map[uint64]*common.BuilderGetValidatorsResponseEntry)
-	backend.relay.proposerDutiesMap[headSlot+1] = &common.BuilderGetValidatorsResponseEntry{
-		Slot: headSlot,
-		Entry: &types.SignedValidatorRegistration{
-			Message: &types.RegisterValidatorRequestMessage{
-				Pubkey:       boosttypes.PublicKey(proposerPubkey),
-				FeeRecipient: feeRec,
-			},
-		},
-	}
-	backend.relay.payloadAttributes = make(map[string]payloadAttributesHelper)
-	backend.relay.payloadAttributes[parentHash] = payloadAttributesHelper{
-		slot:       submissionSlot,
-		parentHash: parentHash,
-		payloadAttributes: beaconclient.PayloadAttributes{
-			PrevRandao: prevRandao,
-		},
-		withdrawalsRoot: phase0.Root(withdrawalsRoot),
-	}
-
-	// Test 1: Try adding txs with higher value
-	req := new(common.TobTxsSubmitRequest)
-	addr1 := common2.HexToAddress("0xB9D7a3554F221B34f49d7d3C61375E603aFb699e")
-
-	tx3 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    3,
-		GasPrice: big.NewInt(3),
-		Gas:      3,
-		To:       &addr1,
-		Value:    big.NewInt(3),
-		Data:     []byte("tx3"),
-	})
-	tx4 := gethtypes.NewTx(&gethtypes.LegacyTx{
-		Nonce:    4,
-		GasPrice: big.NewInt(5),
-		Gas:      12,
-		To:       &backend.relay.relayerPayoutAddress,
-		Value:    big.NewInt(10),
-		Data:     []byte(""),
-	})
-	tx3byte, err := tx3.MarshalBinary()
-	require.NoError(t, err)
-	tx4byte, err := tx4.MarshalBinary()
-	require.NoError(t, err)
-	txs := bellatrixUtil.ExecutionPayloadTransactions{Transactions: []bellatrix.Transaction{tx3byte, tx4byte}}
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot + 1,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err := req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr := backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	tobTxValue, err := backend.redis.GetTobTxValue(context.Background(), backend.redis.NewPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-	fmt.Printf("tobTxValue in test is : %v\n", tobTxValue)
-	require.Equal(t, tobTxValue, big.NewInt(10))
-
-	tobtxs, err := backend.redis.GetTobTx(context.Background(), backend.redis.NewTxPipeline(), headSlot+1, parentHash)
-	require.NoError(t, err)
-
-	require.Equal(t, 2, len(tobtxs))
-
-	firstTx := new(gethtypes.Transaction)
-	err = firstTx.UnmarshalBinary(tobtxs[0])
-
-	firstTxJson, err := firstTx.MarshalJSON()
-	require.NoError(t, err)
-	tx1Json, err := tx3.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, firstTxJson, tx1Json)
-
-	secondTx := new(gethtypes.Transaction)
-	err = secondTx.UnmarshalBinary(tobtxs[1])
-	secondTxJson, err := secondTx.MarshalJSON()
-	require.NoError(t, err)
-	tx2Json, err := tx4.MarshalJSON()
-	require.NoError(t, err)
-	require.Equal(t, secondTxJson, tx2Json)
-
-	// Test 4: Past slot
-	req = &common.TobTxsSubmitRequest{
-		ParentHash: parentHash,
-		TobTxs:     txs,
-		Slot:       headSlot - 2,
-	}
-	fmt.Printf("Marshalling request to json!!")
-	jsonReq, err = req.MarshalJSON()
-	require.NoError(t, err)
-	fmt.Printf("Unmarshalling request from json!!")
-
-	rr = backend.requestBytes(http.MethodPost, path, jsonReq, map[string]string{
-		"Content-Type": "application/json",
-	})
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Contains(t, rr.Body.String(), "Submitted TOB tx request for past slot!")
 }
